Fix handling for broken .eml messages.
I didn't test this well enough yesterday. It turns out that the
attachment was getting created, but its ContentType was overwritten to
message/rfc822. This caused problems elsewhere because the the
attachment we created did not have the structure expected for
message/rfc822.

The fix: Don't let Body override our ContentType declaration.