Specify tabmodule when calling yacc to avoid regenerating parsetab.py
Yacc tries to load the previously generated parsetab.py, but because
we're building a bunch of different parsers with different start tokens
we always fail to load the cached parsetab.py because the signatures
don't match. This diff changes things so that we generate a unique
parsetab.py for each of the different parsers. We also commit these
parsetabs so they don't need to be regenerated at startup when flanker is
imported.