Expose __version__ module-level attribute
It is possible for integrators to need to check the flanker version
being run by the client in order to enable/disable flanker
functionality depending on which python version is being run.

Indeed, if the python version is 2.x, any version of flanker should
work, however if the version is 3.x, only flanker versions >= 0.9 are
compatible.

This patch makes flanker comply with PEP396 [1] and allows for runtime
version checking in cases where a hard dependency requiring flanker >=
0.9 is not possible (e.g. optional e-mail validation)

[1] https://www.python.org/dev/peps/pep-0396/#specification